@@ -13,15 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/cdk": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/localize": "^9.1.13",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@auth0/angular-jwt": "^4.0.0",
     "@fortawesome/angular-fontawesome": "^0.3.0",
     "@fortawesome/fontawesome-free": "^5.15.3",
@@ -31,14 +32,16 @@
     "@types/chart.js": "^2.9.3",
     "@types/heatmap.js": "^2.0.36",
     "@types/leaflet": "^1.7.0",
+    "@types/leaflet.heat": "^0.2.0",
     "animate.css": "^4.1.0",
     "chart.js": "^2.9.4",
     "chartjs-plugin-datalabels": "^1.0.0",
     "core-js": "^2.5.4",
-    "heatmap.js": "^2.0.5",
+    "heatmap.js": "git+https://github.com/ionata/heatmap.js.git",
     "jwt-decode": "^2.2.0",
     "leaflet": "^1.7.1",
     "leaflet-heatmap": "^1.0.0",
+    "leaflet.heat": "https://github.com/Leaflet/Leaflet.heat/tarball/gh-pages",
     "lodash": "^4.17.21",
     "moment": "^2.27.0",
     "moment-timezone": "^0.5.33",
@@ -52,14 +55,14 @@
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "^0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "^2.8.18",
     "@types/jasminewd2": "^2.0.10",
-    "@types/lodash": "^4.14.177",
     "@types/node": "~8.9.4",
+    "@types/geojson": "7946.0.8",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -72,6 +75,6 @@
     "sass": "^1.49.5",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   }
 }
